+++ resources/views/rooms/create.blade.php
@@ -65,13 +65,9 @@
         </div>
     </div>
 
-    <!-- JavaScriptの読み込み -->
-    @vite('resources/js/debate-form.js')
     <script>
-        // 設定とデータ
-        const debateFormConfig = {
-            formType: 'room',
-            formSelector: '#room-create-form',
+        // 設定データをJavaScriptページエントリーポイントに渡す
+        window.roomCreateConfig = {
             formats: @json($translatedFormats),
             requiredFields: [
                 { name: 'topic', message: "{{ __('messages.topic_required') }}" },
@@ -105,15 +101,6 @@
                 fieldRequired: "{{ __('messages.field_required') }}"
             }
         };
-
-        // ディベートフォームマネージャーの初期化
-        document.addEventListener('DOMContentLoaded', function() {
-            const debateForm = new DebateFormManager(debateFormConfig);
-            debateForm.init();
-
-            // グローバル参照の設定（後方互換性のため）
-            window.stepManager = debateForm.stepManager;
-            window.formatManager = debateForm.formatManager;
-        });
     </script>
+    @vite(['resources/js/pages/room-create.js'])
 </x-app-layout>
